chore(build): enable Wextra
Annotate unused retail callback parameters and make warning-only conversions explicit.

Model thread collision callbacks as raw code pointers with call-site-specific signatures, matching the retail funcThCollide slot while keeping GCC -Wextra clean.

+++ game/233/CS_Instance.c
@@ -152,7 +152,7 @@ int CS_Instance_GetNumAnimFrames(struct Instance *modelInst, int animIndex, int 
 		return 0;
 	}
 
-	if (animIndex >= header->numAnimations)
+	if (animIndex >= (int)header->numAnimations)
 	{
 		return 0;
 	}
